+++ eslint.config.mjs
@@ -21,7 +21,7 @@ export default [
 				...globals.mocha,
 			},
 
-			ecmaVersion: 9,
+			ecmaVersion: 11,
 			sourceType: "module",
 		},