+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 	- `hideEmpty` - set to `true` if you can hide an empty ToC
 	- `sortBy` -  `'page'` (default) or `'title'`
 	- `sortLocale` - custom locale to sort by property `sortBy`
+- Fixed margin override with 0 value
 
 ## 0.3.2 - 2026-01-11