+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added SVG validation: width and height must be specified (in SVG string/element or `svg` node)
 - Added support for image scaling with only `height` defined
 - Added custom `markerColor` for each item of `ul` and `ol`
+- Fixed svg-to-pdfkit - TypeError: t.classList.contains is not a function
 
 ## 0.3.1 - 2026-01-07